+++ forui/lib/src/widgets/select/multi/multi_select.dart
@@ -41,7 +41,7 @@ abstract class FMultiSelect<T> extends StatelessWidget {
 
   /// The default loading builder that shows a spinner when an asynchronous search is pending.
   static Widget defaultSearchLoadingBuilder(BuildContext _, FSelectSearchStyle style, Widget? _) => Padding(
-    padding: const EdgeInsets.all(8.0),
+    padding: const EdgeInsets.all(13),
     child: FProgress.circularIcon(style: (s) => style.loadingIndicatorStyle),
   );
 
@@ -49,7 +49,7 @@ abstract class FMultiSelect<T> extends StatelessWidget {
   static Widget defaultEmptyBuilder(BuildContext context, FMultiSelectStyle style, Widget? _) {
     final localizations = FLocalizations.of(context) ?? FDefaultLocalizations();
     return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 15),
+      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 14),
       child: Text(localizations.selectNoResults, style: style.emptyTextStyle),
     );
   }